requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
click==8.1.7
colorama==0.4.6
fastapi==0.115.6
frozenlist==1.5.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
pytest-asyncio==0.24.0
python-multipart==0.0.19
PyYAML==6.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
yarl==1.18.3
